extracted sf-86 fields
